+++ Solutions/Tenant Information Portal/src/TIP.Common/Diagnostics/EventCategory.cs
@@ -1,0 +1,19 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace TIP.Common.Diagnostics
+{
+    public enum EventCategory
+    {
+        Unknown,
+        Cache,
+        Database,
+        Page,
+        Service,
+        API,
+        Authorization
+    }
+}