+++ Solutions/Tenant Information Portal/src/TIP.DashBoard/App_Start/FilterConfig.cs
@@ -20,7 +20,6 @@
 //SOFTWARE.
 // ------------------------------------------------------------------------------
 
-
 using System.Web.Mvc;
 
 namespace TIP.Dashboard